Simplify Arrow type and methods dealing with them.
Removed HandArrow and SquareArrow interfaces, combining them into one interface named Arrow.
Removed drawHandArrow() and drawSqaureArrow() methods of the GUI class, replacing with a drawSnapArrow() method.
drawSnapArrow() method draws an arrow that snaps to key points on the canvas(squares and hand pieces)

To check if an arrow is a SquareArrow, check if arrow.src is a string. And if it's not, then it's a HandArrow.